api/requirements.txt
tensorflow==2.17.0
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4